refactor: divide onboarding content in separate components
